Move reduxForm into <StageEditor />
